requirements/base.txt
Django>=5.1.1,<5.2
djangorestframework>=3.15.2,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4
drf-yasg>=1.21.7,<1.22
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0.1,<1.2



